backend/requirements.txt
fastapi==0.113.0
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.15.2
pillow==10.0.0
pydantic==2.10.5
python-multipart==0.0.6
